% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theming.R
\docType{data}
\name{bms_theme}
\alias{bms_theme}
\title{commonly used ggplot theme customization}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
bms_theme
}
\description{
commonly used ggplot theme customization
}
\keyword{datasets}
