% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cipher.R
\name{ciper}
\alias{ciper}
\title{cipher
Calls base::chartr to scramble input based on a set key and seed.
For amino acids, single letter codes are: GPAVLIMCFYWHKRQNEDST}
\usage{
ciper(input, seed, key = c(LETTERS))
}
\arguments{
\item{input}{Character vector of strings to encode.}

\item{seed}{Parameter to pass to set.seed for reproducibility.}

\item{key}{Poissible characters to found in input and to include in output.}
}
\description{
cipher
Calls base::chartr to scramble input based on a set key and seed.
For amino acids, single letter codes are: GPAVLIMCFYWHKRQNEDST
}
