% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_df_reps.R
\name{collapse_df_reps}
\alias{collapse_df_reps}
\title{collapse_df_reps}
\usage{
collapse_df_reps(lhs, rhs)
}
\arguments{
\item{df}{data frame or tibble with observations as columns}

\item{pid}{patient id, character vector with length equal to number of columns in df. replicate columns are indicated by same value in corresponding elements of pid}

\item{method}{what kind of collapsing to do. min = take minimum value. mean = take mean of values. clean = take mean, but if either value is below clean_thresh, set final valut o zero. cor: compute correlation across all features for data. required duplicates.}

\item{clean_thresh}{for method="clean". threshold value below which if any replicate exhibits this value, set output to 0}

\item{ignore_cols}{columns in df to exclude from colapse (e.g. feature id)}

\item{show_progress}{logical indicating whether or not to show progress bar}
}
\description{
collapse columns in df which chare an identity index provided in the vector pid
}
