% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_bins.R
\name{compute_breadth}
\alias{compute_breadth}
\title{compute_breadth}
\usage{
compute_breadth(df, ignore_cols = c("id", "bin"))
}
\arguments{
\item{df}{data frame of enrichment scores}

\item{ignore_cols}{column of data frame containing metadata that chould be excluded from breadth calculation}
}
\description{
count number of hits per sample
}
