% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyTable.R
\name{copyTable}
\alias{copyTable}
\title{copyTable}
\usage{
copyTable(df, sep = "\\t", row.names = FALSE, ...)
}
\arguments{
\item{df}{Data frame to copy to clipboard.}

\item{sep}{Separator}

\item{row.names}{Logical whether or not to include row names}

\item{...}{Additional parameters passed to write.table}
}
\description{
write data frame to clipboard
}
