% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{crossing}
\alias{crossing}
\title{useful functions for other packages that i should remember to use more often}
\usage{
crossing(...)
}
\description{
tidyr::crossing is like base::expand.grid. all unique combinations of elements from 2 vectors
}
