% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deFactor.R
\name{deFactor}
\alias{deFactor}
\title{lapply as.character to data frame columns to convert factor to character}
\usage{
deFactor(df)
}
\arguments{
\item{df}{Data frame to convert.}
}
\description{
lapply as.character to data frame columns to convert factor to character
}
