% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchange.R
\name{foldchange}
\alias{foldchange}
\title{foldchange
Calculate fold change from rpm scaled data.}
\usage{
foldchange(rpm_control, rpm_sample, binsize = 300, ignore_cols = c())
}
\arguments{
\item{rpm_control}{Reads per million (counts per million) data frame of control data from which to determine bins (e.g. beads only). Peptide order should match that of rpm_sample.}

\item{rpm_sample}{Reads per million (counts per million) data frame of sample data for which to calculate z-scores. Peptide order should match that of rpm_control}

\item{binsize}{Number of peptides per bin (default 300)}

\item{ignore_cols}{Column names to exclude from calculation (e.g. id column)}
}
\description{
foldchange
Calculate fold change from rpm scaled data.
}
