% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{fr}
\alias{fr}
\title{functions for reading and writing data}
\usage{
fr(file, ...)
}
\arguments{
\item{file}{directory path to table}

\item{...}{additional parameters to pass to data.table::fread}
}
\description{
fread fwrite wrappers
}
\details{
fr
}
