% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{fw}
\alias{fw}
\title{functions for reading and writing data}
\usage{
fw(x, file, ...)
}
\arguments{
\item{x}{table to write}

\item{file}{directory path destination to write}

\item{...}{additional parameters to pass to data.table::fwrite}
}
\description{
fread fwrite wrappers
fw
}
