% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapExtract.R
\name{heatmapExtract}
\alias{heatmapExtract}
\title{Input jpg heatmap, color scale, table dimensions. Output matrix of corresponding values.}
\usage{
heatmapExtract(heatmap, scale, scale_range = c(1, 0),
  scale_direction = "vertical", table_dim, n_samples = 1,
  verbose = TRUE)
}
\arguments{
\item{heatmap}{Either image array or path to cropped jpg containing heatmap color info}

\item{scale}{Either image array or path to cropped jpg containing color scale info}

\item{scale_range}{Numeric vector of length two containing in order the topmost and bottommost (or leftmost and rightmost) value in the color scale.}

\item{scale_direction}{Character vector, "vertical" or "horizontal" describing color scale orientation.}

\item{table_dim}{Numeric vector of length two containing in order the number of rows and number of columns in the heatmap image}

\item{n_samples}{Number of pixels to sample per cell. By default, heatmapExtract estimates the central pixel of each cell based on the heatmap pixel dimensions and table_dim. If n_samples is set > 1, then heatmapExtract will average across multiple near-central pixels.}

\item{verbose}{Logical, if true then debugging information will be printed.}
}
\description{
Input jpg heatmap, color scale, table dimensions. Output matrix of corresponding values.
}
