% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_bins.R
\name{hits_score}
\alias{hits_score}
\title{hits_score}
\usage{
hits_score(df, z_thresh = 5, ignore_cols = c("id", "bin"))
}
\arguments{
\item{df}{data frame of z-scores}

\item{z_thresh}{threshold above which to consider a hit.}

\item{ignore_cols}{columns to ignore from hit analysis (e.g. feature id)}
}
\description{
compute hits based on z-scores, value abot threhsold
}
