% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pro_fun.R
\name{new_pro_fun}
\alias{new_pro_fun}
\title{new_pro_fun}
\usage{
new_pro_fun(
  zscore_df,
  protein_split,
  method = c("sum", "max", "breadth"),
  z_thresh = 5
)
}
\arguments{
\item{zscore_df}{data frame of zscores}

\item{protein_split}{character vector ewqual to number of rows in zscore df indicating which peptides belong to which proteins}

\item{method}{whether to sum values across peptides, max, or take number of peptides above a threshold}

\item{z_thresh}{threhold above which to count as scoring for method="breadth}
}
\description{
calculate prosum, promax, pro_breadth -- like polyclonal but not redundancy collapsed
}
