% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteTable.R
\name{pasteTable}
\alias{pasteTable}
\title{pasteTable}
\usage{
pasteTable(sep = "\\t", header = FALSE, ...)
}
\arguments{
\item{sep}{table separator value. e.g. tab, space, comma, etc.}

\item{header}{whether or not to assign the first row to column names. default FALSE}

\item{...}{Additional parameters passed to utils::read.table}
}
\description{
Function to move table from clipboard to R
}
