% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteWinpath.R
\name{pasteWinpath}
\alias{pasteWinpath}
\title{pasteWinpath gsub wrapper. By default changes \ to / from windows paths from cliboard.}
\usage{
pasteWinpath(sub = "/")
}
\arguments{
\item{sub}{Optional gsub to replace '\\' with in converted path. Set to '\\\\' to ignore.}
}
\description{
pasteWinpath gsub wrapper. By default changes \ to / from windows paths from cliboard.
}
