% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{read_csv}
\alias{read_csv}
\title{functions for reading and writing data}
\usage{
read_csv(file, ...)
}
\arguments{
\item{file}{directory path to csv}

\item{...}{additional parameters to pass to readr::read_csv}
}
\description{
fread fwrite wrappers
}
\details{
read_csv
}
