% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncipher.R
\name{uncipher}
\alias{uncipher}
\title{uncipher
Calls base::chartr to unscramble input based on a set key and seed.
For amino acids, single letter codes are: GPAVLIMCFYWHKRQNEDST}
\usage{
uncipher(input, seed, key = c(LETTERS))
}
\arguments{
\item{input}{Character vector of strings to encode.}

\item{seed}{Parameter to pass to set.seed for reproducibility.}

\item{key}{Poissible characters to found in input and to include in output.}
}
\description{
uncipher
Calls base::chartr to unscramble input based on a set key and seed.
For amino acids, single letter codes are: GPAVLIMCFYWHKRQNEDST
}
