% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vir_evidence.R
\name{vir_evidence}
\alias{vir_evidence}
\title{vir_evidence
Consolidate virscan hits to evidence peptides. Adapted from https://pubmed.ncbi.nlm.nih.gov/26045439/
Group hit peptides by species. Rank species by number of hits.
Starting with species with most hits, for each hit peptide, remove other hit peptides with 7aa overlap
from lower hit-breadth species.}
\usage{
vir_evidence(
  hit_ids,
  hit_species,
  alignment_matrix = NULL,
  hit_aa = NULL,
  similarity_method = c("overlap", "blast"),
  similarity_thresh = 7
)
}
\arguments{
\item{hit_ids}{vector of peptide ids enriched for sample}

\item{hit_species}{vector of viral species ids corresponding to hit_ids}

\item{alignment_matrix}{binary matrix indicating whether two peptides are similar (e.g. share 7aa ovlerap or blast alignment). matrix rownames and column names should correspond to peptide ids used in hit_ids. use either alignment_matrix or hit_aa}

\item{hit_aa}{amino acid sequences of peptides listed by hit_ids. use either hit_aa or alignment_matrix.}

\item{similarity_method}{method to use for characterizing similarity between peptide seuqneces provided by hit_aa.}

\item{similarity_thresh}{amino acid overlap threshold or blast e-value threshold to use for similarity cutoff.}
}
\description{
vir_evidence
Consolidate virscan hits to evidence peptides. Adapted from https://pubmed.ncbi.nlm.nih.gov/26045439/
Group hit peptides by species. Rank species by number of hits.
Starting with species with most hits, for each hit peptide, remove other hit peptides with 7aa overlap
from lower hit-breadth species.
}
