% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_bins.R
\name{zscore_bins}
\alias{zscore_bins}
\title{zscore_bins
Calculate zscores from rpm scaled data. Adapted from Michael Mina et al. 2019 DOI: 10.1126/science.aay6485}
\usage{
zscore_bins(rpm_control, rpm_sample, binsize = 300,
  include_percentile = c(0.05, 0.95))
}
\arguments{
\item{rpm_control}{Reads per million (counts per million) data frame of control data from which to determine bins (e.g. beads only). Peptide order should match that of rpm_sample.}

\item{rpm_sample}{Reads per million (counts per million) data frame of sample data for which to calculate z-scores. Peptide order should match that of rpm_control}

\item{binsize}{Number of peptides per bin (default 300)}

\item{include_percentile}{Peptide percentiles range to include in mean and sd calculation (default 0.05, 0.95)}
}
\description{
zscore_bins
Calculate zscores from rpm scaled data. Adapted from Michael Mina et al. 2019 DOI: 10.1126/science.aay6485
}
