% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyWinpath.R
\name{copyWinpath}
\alias{copyWinpath}
\title{copyWinpath}
\usage{
copyWinpath(sub = "/")
}
\arguments{
\item{sub}{Optional gsub to replace '\\' with in converted path. Set to '\\\\' to ignore.}
}
\description{
write converted windows path (from \ to /) to clipboard
}
