% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_bins.R
\name{zscore_bins}
\alias{zscore_bins}
\title{zscore_bins
Calculate zscores from rpm scaled data. Adapted from Michael Mina et al. 2019 DOI: 10.1126/science.aay6485}
\usage{
zscore_bins(
  rpm_control,
  rpm_sample,
  binsize = 300,
  include_percentile = c(0.05, 0.95),
  ignore_cols = c(),
  export_bins = TRUE,
  sd_min = NULL
)
}
\arguments{
\item{rpm_control}{Reads per million (counts per million) data frame of control data from which to determine bins (e.g. beads only). Peptide order should match that of rpm_sample.}

\item{rpm_sample}{Reads per million (counts per million) data frame of sample data for which to calculate z-scores. Peptide order should match that of rpm_control}

\item{binsize}{Number of peptides per bin (default 300)}

\item{include_percentile}{Peptide percentiles range to include in mean and sd calculation (default 0.05, 0.95)}

\item{ignore_cols}{Column names to exclude from calculation (e.g. id column)}

\item{export_bins}{Logical. if true, include a column called ranks indicating bins based on rpm_control values}

\item{sd_min.}{Default value NULL. If set to a numeric value or vector, can be used to impose a floor to the standard deviation. If a single value is given, the same floor is applied to the standard deviation of each bin. If a vector is provided, each value is applied sequentially.}
}
\description{
zscore_bins
Calculate zscores from rpm scaled data. Adapted from Michael Mina et al. 2019 DOI: 10.1126/science.aay6485
}
